backend/requirements.txt
Django>=4.2
dj-database-url>=2.2
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
djangorestframework-simplejwt>=5.3
channels>=4.0
channels-redis>=4.1
django-jazzmin>=3.0
psycopg2-binary>=2.9
whitenoise>=6.6
gunicorn>=21.2
